Move StopSearchResult result mappers to generic utils
