+++ ui/src/components/stop-registry/search/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './handleAllMunicipalities';
+export * from './resultMappers';
 export * from './useNumericSortingCollator';
 export * from './useStopSearchUrlState';
